Moves isGambitCampaignsTemplate to helpers/templates
