+++ config/lib/helpers.js
@@ -11,21 +11,6 @@ module.exports = {
     'askSignup',
     'invalidAskSignupResponse',
   ],
-  gambitCampaignsTemplates: [
-    'askCaption',
-    'askPhoto',
-    'askQuantity',
-    'askWhyParticipated',
-    'completedMenu',
-    'externalSignupMenu',
-    'gambitSignupMenu',
-    'invalidCaption',
-    'invalidCompletedMenuCommand',
-    'invalidPhoto',
-    'invalidQuantity',
-    'invalidSignupMenuCommand',
-    'invalidWhyParticipated',
-  ],
   menuCommand,
   macros: {
     confirmedCampaign: 'confirmedCampaign',
