+++ config/lib/helpers/templates.js
@@ -5,7 +5,24 @@ const helpCenterUrl = 'http://doso.me/1jf4/291kep';
 // @see lib/helpers.js
 const menuCommand = 'menu';
 
+const gambitCampaignsTemplates = [
+  'askCaption',
+  'askPhoto',
+  'askQuantity',
+  'askWhyParticipated',
+  'completedMenu',
+  'externalSignupMenu',
+  'gambitSignupMenu',
+  'invalidCaption',
+  'invalidCompletedMenuCommand',
+  'invalidPhoto',
+  'invalidQuantity',
+  'invalidSignupMenuCommand',
+  'invalidWhyParticipated',
+];
+
 module.exports = {
+  gambitCampaignsTemplates,
   templateText: {
     badWords: 'Not cool. I\'m a real person & that offends me. I send out these texts to help young ppl take action. If you don\'t want my texts, text STOP or LESS to get less.',
     crisis: 'Thanks for being brave and sharing that. If you want to talk to someone, our friends at CTL are here for you 24/7. Just send a text to 741741. Theyll listen!',
