+++ lib/helpers.js
@@ -30,13 +30,6 @@ module.exports.isDeclinedCampaignMacro = function (text) {
   return (text === config.macros.declinedCampaign);
 };
 
-module.exports.isGambitCampaignsTemplate = function (templateName) {
-  const result = config.gambitCampaignsTemplates.includes(templateName);
-  logger.debug('isGambitCampaignsTemplate', { templateName, result });
-
-  return result;
-};
-
 module.exports.isMacro = function (text) {
   const result = config.macros[text];
   logger.debug('isMacro', { text, result });
